// src/types/apps/restaurantTypes.ts
export type RestaurantTypes = {
  id: number
  name: string
  address: string
  phoneNumber: string
  email: string
  website?: string
  vatNumber?: string
  posId?: string
  rel?: string
  terminal?: string
  pluHash?: string
  ticketTeller?: string
  ticketSignature?: string
  controlModuleId?: string
  vatCardId?: string
  description: string
  status: string
}